Fix memory leaks in strategic pathing
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@1536 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ SCREENS.cpp
@@ -19,11 +19,11 @@ int iScreenMode = 0;	/* 0 for Fullscreen */
 #endif
 BOOL bScreenModeCmdLine = FALSE;
 
-int SCREEN_WIDTH;
-int SCREEN_HEIGHT;
+UINT16 SCREEN_WIDTH;
+UINT16 SCREEN_HEIGHT;
 
-int iScreenWidthOffset;
-int iScreenHeightOffset;
+INT16 iScreenWidthOffset;
+INT16 iScreenHeightOffset;
 
 BOOLEAN fDisplayOverheadMap;
